% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{Loss}
\alias{Loss}
\title{Matrix of Value at Risk losses coming from 10 ARCH-type models}
\format{a matrix object.}
\usage{
Loss
}
\description{
Matrix of Losses associated to a forecast series of 2000 observation of the VaR calculated at the 1%
confidence level. This is a 2000*10 matrix, the losses are calculated using the Asymmetric Loss function of Gonzales et.al. (2004).
}
\author{
Leopoldo Catania, 2014-07-27
}
\references{
Gonzalez-Rivera G, Lee TH, Mishra S (2004). Forecasting volatility: A reality check based on option pricing, utility function, value-at-risk, and predictive likelihood." International Journal of Forecasting, 20(4), 629-645. ISSN 0169-2070. doi:http://dx.doi.org/10.1016/j.ijforecast.2003.10.003. URL http://www.sciencedirect.com/science/article/pii/S0169207003001420
}

