% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LossLevel}
\alias{LossLevel}
\title{Loss Function for level forecasts}
\usage{
LossLevel(realized, evaluated, which = "SE")
}
\arguments{
\item{realized}{a vector with the realizations of the interest object.}

\item{evaluated}{a vector or a matrix of forecasts}

\item{which}{The loss function to use. possible choices are: \code{'SE'} that coincides with Square Error and \code{AE} that coincides
with Absolute Error}
}
\value{
A matrix with the forecast losses
}
\description{
Calculate the losses associated with level forecasts
}
\author{
Leopoldo Catania & Mauro Bernardi
}

