\name{MAplot}
\alias{MAplot}
\title{Moving Average plot}
\usage{
MAplot(date,price,num_ma)
}
\description{
Calculate the moving average of data, then plot the data and line.
}
\arguments{
  \item{date}{the date of price}
  \item{price}{daily price}
  \item{num_ma}{the time you want to build}
}
