\name{RR}
\alias{RR}
\title{Rate of Return}
\usage{
RR(data,type = "day")
}
\description{
This function will help you to calculate the rate of return.
Input your daily data,it will output the rate.
}
\arguments{
  \item{data}{daily price}
  \item{type}{for day,week,month,season,year}
}
