% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{STOXXIndexesRet}
\alias{STOXXIndexesRet}
\title{STOXX indexes logarithmic returns from 1992-01-02 to 2014-07-24}
\format{a \code{xts} object.}
\source{
www.stoxx.com
}
\usage{
STOXXIndexesRet
}
\description{
Daily logarithmic returns of the STOXX North America 600 (SXA1E) the
STOXX Asia/Pacific 600 (SXP1E) the STOXX Europe 600 (SXXP) and the STOXX Global 1800 (SXW1E) from 1992-01-02 to 2014-07-24.
}
\author{
Leopoldo Catania, 2014-07-27
}

